[refactor] US data now has its own store
US data was stored in the same json file as other countries data;
but being very big (> 50MB), it would slow down running for other
countries as soon as it was downloaded.

To avoid this, we store US data in a dedicated file, so we don't
have to parse it when trying to access another country's data.